requirements.txt
tensorflow==2.14.0
streamlit-audiorec==0.1.3
matplotlib==3.8.1
librosa==0.10.1
numpy==1.26.1
soundfile==0.12.1
